lambda/generate-presigned-url/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
attrs==25.3.0
    # via flake8-bugbear
black==25.1.0
    # via -r requirements-dev.in
boto3-stubs[s3]==1.39.3
    # via -r requirements-dev.in
botocore-stubs==1.38.46
    # via boto3-stubs
cfgv==3.4.0
    # via pre-commit
click==8.2.1
    # via black
distlib==0.3.9
    # via virtualenv
filelock==3.18.0
    # via virtualenv
flake8==7.3.0
    # via
    #   -r requirements-dev.in
    #   flake8-bugbear
    #   flake8-comprehensions
flake8-bugbear==24.12.12
    # via -r requirements-dev.in
flake8-comprehensions==3.16.0
    # via -r requirements-dev.in
flake8-import-order==0.19.2
    # via -r requirements-dev.in
identify==2.6.12
    # via pre-commit
mccabe==0.7.0
    # via flake8
mypy-boto3-s3==1.39.2
    # via boto3-stubs
mypy-extensions==1.1.0
    # via black
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via black
pathspec==0.12.1
    # via black
platformdirs==4.3.8
    # via
    #   black
    #   virtualenv
pre-commit==4.2.0
    # via -r requirements-dev.in
pycodestyle==2.14.0
    # via
    #   flake8
    #   flake8-import-order
pyflakes==3.4.0
    # via flake8
pyyaml==6.0.2
    # via pre-commit
types-awscrt==0.27.4
    # via botocore-stubs
types-s3transfer==0.13.0
    # via boto3-stubs
virtualenv==20.31.2
    # via pre-commit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
